==> submission2/stacking_output/2. Run_papermill.ipynb <==
# Running the same notebook with multiple input parameters

import papermill as pm
import random
import os

for i in range(1,11):
    pm.execute_notebook(
        'STACKING.ipynb',           # Template Notebook
        f'stacking_file_final{i:02}.ipynb',   # Template를 사용해 Output이 저장되는 Notebook
        parameters = dict(                         # Parameter 설정
            VERSION = f'9.{i:02}',
            NFOLDS = random.randint(5,7),
            SEED = random.randint(0,10000),
        )      
    )
#    os.remove(f'Competition_CatBoost_v3_{i:02}.ipynb')

# for i in range(1,4):
#     pm.execute_notebook(
#         'power_mean_ensemble.ipynb',          
#         f'power_mean_ensemble_{i:02}.ipynb',  
#         parameters = dict(                   
#             P = i,
#         )      
#     )
#     os.remove(f'power_mean_ensemble_{i:02}.ipynb')

# End
